+++ src/ts/static/linksIcons.ts
@@ -15,6 +15,14 @@ const linksIcons = [
         linkTitle: 'Special Thanks',
         iconName: 'thanks.png',
     },
+    {
+        linkTitle: '@official_php',
+        iconName: 'x.svg',
+    },
+    {
+        linkTitle: '@php@fosstodon.org',
+        iconName: 'mastodon.png',
+    },
 ]
 
 export default linksIcons