+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Changed logic with loading recommended videos on the home page
 - Added title for Recommended videos and Recommended links
 - Rewrote the whole logic related to the links section on the home page sidebar. Now it's more structured and easier to maintain. Now, we parse links first and then render them in a proper DOM structure with some additional links like PHP Sandbox, Github, Docker image, etc.
+- Added "Social Links" section to the home page sidebar
 
 ## v2.8.0 (2024-04-29)
 - Aligned the "Show Comments" button vertically on the function page